改进：精简部分代码，移除冗余的变量初始化，精简装箱/拆箱，

+++ core/src/main/java/com/alibaba/fastjson2/JSONWriterUTF8.java
@@ -2029,7 +2029,7 @@ class JSONWriterUTF8
                 continue;
             }
 
-            long v = item.longValue();
+            long v = item;
             boolean writeAsString = nonStringAsString
                     || (browserCompatible && v <= 9007199254740991L && v >= -9007199254740991L);
             if (writeAsString) {
